requirements.txt
openai-whisper
fastapi
uvicorn
notion-client
jieba
python-dotenv
requests
flask
opencc-python-reimplemented